+++ examples/hobotnica/data.py
@@ -1,0 +1,15 @@
+GITHUB_USERNAME = "github-user"
+GITHUB_PERSONAL_TOKEN = "github-user-personal-token"
+GITHUB_REPOSITORY = "github-repo"
+
+GITHUB_REPOSITORIES = {
+    GITHUB_USERNAME: {
+        GITHUB_REPOSITORY: {
+            "uid": "ac711a94-adfa-43ff-8f11-195d78c6f57d",
+            "owner": GITHUB_USERNAME,
+            "name": GITHUB_REPOSITORY,
+            "jobs": ["test", "deploy"],
+            "status": "ready",
+        }
+    }
+}