+++ rororo/openapi/constants.py
@@ -7,5 +7,11 @@ OPENAPI_IS_VALIDATE_RESPONSE_APP_KEY = "rororo_openapi_is_validate_response"
 #: Key to store OpenAPI schema within the ``web.Application`` instance
 OPENAPI_SCHEMA_APP_KEY = "rororo_openapi_schema"
 
+#: API Key security scheme
+OPENAPI_SECURITY_API_KEY = "apiKey"
+
+#: HTTP security scheme
+OPENAPI_SECURITY_HTTP = "http"
+
 #: Key to store OpenAPI spec within the ``web.Application`` instance
 OPENAPI_SPEC_APP_KEY = "rororo_openapi_spec"